Implement IBM i support more cleanly
Add separate definitions for IBM i and use that instead of AIX. On
Python 3.9, we no longer claim to be AIX so that fits better anyway.

Probably should standardize on "OS400" or "IBMI", though.

AFAICT, libutil is not needed at all and /QOpenSys/pkgs/include is
already standard on any recent GCC build, so those do not need to be
added in setup.py either.